+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 **July 24, 2023 [DESCW-`1331`](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1331)**
 - Frontend
-  - Addedfix that removed parameters if you change type.
+  - Added fix that removed parameters if you change type.
+  - Added react router protected route to report routes
 
 **July 24, 2023 [DESCW-1330](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1330)**
 - Backend